Add team member blocks: contact email, contact number, gallery, role, social links, and tagline
