+++ src/blocks/role/index.js
@@ -1,0 +1,70 @@
+/**
+ * Team Role Block Variation
+ *
+ * @since 2.2.0
+ * @package TO_Team
+ */
+
+import { __ } from '@wordpress/i18n';
+import { registerForPostTypesAndTemplates } from '@utils/conditional-block-registration.js';
+
+wp.domReady(() => {
+    const registerTeamRoleVariation = () => {
+        wp.blocks.registerBlockVariation('core/group', {
+            name: 'lsx-tour-operator/team-role',
+            title: __('Team - Role', 'to-team'),
+            icon: 'businessman',
+            category: 'lsx-tour-operator',
+            description: __('Display the role for this team member.', 'to-team'),
+            keywords: [__('role', 'to-team'), __('team', 'to-team'), __('position', 'to-team'), __('job', 'to-team')],
+            isActive: (blockAttributes, variationAttributes) => {
+                return blockAttributes.className === variationAttributes.className;
+            },
+            attributes: {
+                metadata: { name: __('Team - Role', 'to-team') },
+                className: 'lsx-role-wrapper',
+                layout: { type: 'flex', flexWrap: 'nowrap', verticalAlignment: 'top' },
+            },
+            innerBlocks: [
+                [
+                    'core/group',
+                    { layout: { type: 'flex', flexWrap: 'nowrap', verticalAlignment: 'middle' } },
+                    [['lsx-tour-operator/icons', { iconType: 'solid', iconName: 'teamIcon' }]],
+                ],
+                [
+                    'core/group',
+                    { layout: { type: 'flex', flexWrap: 'nowrap' } },
+                    [
+                        [
+                            'core/paragraph',
+                            {
+                                metadata: {
+                                    bindings: {
+                                        content: { source: 'lsx/post-meta', args: { key: 'role' } },
+                                    },
+                                },
+                                prefix: __('Role:', 'to-team'),
+                                prefixBold: true,
+                            },
+                        ],
+                    ],
+                ],
+            ],
+            example: {
+                innerBlocks: [
+                    {
+                        name: 'core/group',
+                        attributes: { layout: { type: 'flex', flexWrap: 'nowrap', verticalAlignment: 'middle' } },
+                        innerBlocks: [
+                            { name: 'lsx-tour-operator/icons', attributes: { iconType: 'solid', iconName: 'teamIcon' } },
+                            { name: 'core/paragraph', attributes: { content: '<strong>' + __('Role: ', 'to-team') + '</strong>' + __('Senior Safari Guide', 'to-team') } },
+                        ],
+                    },
+                ],
+            },
+        });
+    };
+
+    const conditionalRegister = registerForPostTypesAndTemplates(['team'], ['team'], registerTeamRoleVariation);
+    conditionalRegister();
+});
