--- src/blocks/team-contact-number/block.json
@@ -1,10 +1,0 @@
-{
-    "$schema": "https://schemas.wp.org/trunk/block.json",
-    "apiVersion": 3,
-    "textdomain": "to-team",
-    "name": "lsx-tour-operator/team-contact-number",
-    "title": "Team Contact Number",
-    "category": "lsx-tour-operator",
-    "description": "Display the contact number for this team member.",
-    "editorScript": "file:index.js"
-}
