--- src/blocks/team-contact-number/index.js
@@ -1,70 +1,0 @@
-/**
- * Team Contact Number Block Variation
- *
- * @since 2.2.0
- * @package TO_Team
- */
-
-import { __ } from '@wordpress/i18n';
-import { registerForPostTypesAndTemplates } from '@utils/conditional-block-registration.js';
-
-wp.domReady(() => {
-    const registerTeamContactNumberVariation = () => {
-        wp.blocks.registerBlockVariation('core/group', {
-            name: 'lsx-tour-operator/team-contact-number',
-            title: __('Team Contact Number', 'to-team'),
-            icon: 'phone',
-            category: 'lsx-tour-operator',
-            description: __('Display the contact number for this team member.', 'to-team'),
-            keywords: [__('phone', 'to-team'), __('contact', 'to-team'), __('team', 'to-team'), __('number', 'to-team')],
-            isActive: (blockAttributes, variationAttributes) => {
-                return blockAttributes.className === variationAttributes.className;
-            },
-            attributes: {
-                metadata: { name: __('Team Contact Number', 'to-team') },
-                className: 'lsx-contact-number-wrapper',
-                layout: { type: 'flex', flexWrap: 'nowrap', verticalAlignment: 'top' },
-            },
-            innerBlocks: [
-                [
-                    'core/group',
-                    { layout: { type: 'flex', flexWrap: 'nowrap', verticalAlignment: 'middle' } },
-                    [['lsx-tour-operator/icons', { iconType: 'solid', iconName: 'phoneIcon' }]],
-                ],
-                [
-                    'core/group',
-                    { layout: { type: 'flex', flexWrap: 'nowrap' } },
-                    [
-                        [
-                            'core/paragraph',
-                            {
-                                metadata: {
-                                    bindings: {
-                                        content: { source: 'lsx/post-meta', args: { key: 'contact_number' } },
-                                    },
-                                },
-                                prefix: __('Phone:', 'to-team'),
-                                prefixBold: true,
-                            },
-                        ],
-                    ],
-                ],
-            ],
-            example: {
-                innerBlocks: [
-                    {
-                        name: 'core/group',
-                        attributes: { layout: { type: 'flex', flexWrap: 'nowrap', verticalAlignment: 'middle' } },
-                        innerBlocks: [
-                            { name: 'lsx-tour-operator/icons', attributes: { iconType: 'solid', iconName: 'phoneIcon' } },
-                            { name: 'core/paragraph', attributes: { content: '<strong>' + __('Phone: ', 'to-team') + '</strong>' + __('+27 21 555 0100', 'to-team') } },
-                        ],
-                    },
-                ],
-            },
-        });
-    };
-
-    const conditionalRegister = registerForPostTypesAndTemplates(['team'], ['team'], registerTeamContactNumberVariation);
-    conditionalRegister();
-});
