--- src/blocks/team-role/block.json
@@ -1,10 +1,0 @@
-{
-    "$schema": "https://schemas.wp.org/trunk/block.json",
-    "apiVersion": 3,
-    "textdomain": "to-team",
-    "name": "lsx-tour-operator/team-role",
-    "title": "Team Role",
-    "category": "lsx-tour-operator",
-    "description": "Display the role for this team member.",
-    "editorScript": "file:index.js"
-}
