--- src/blocks/team-tagline/index.js
@@ -1,70 +1,0 @@
-/**
- * Team Tagline Block Variation
- *
- * @since 2.2.0
- * @package TO_Team
- */
-
-import { __ } from '@wordpress/i18n';
-import { registerForPostTypesAndTemplates } from '@utils/conditional-block-registration.js';
-
-wp.domReady(() => {
-    const registerTeamTaglineVariation = () => {
-        wp.blocks.registerBlockVariation('core/group', {
-            name: 'lsx-tour-operator/team-tagline',
-            title: __('Team Tagline', 'to-team'),
-            icon: 'editor-quote',
-            category: 'lsx-tour-operator',
-            description: __('Display the tagline for this team member.', 'to-team'),
-            keywords: [__('tagline', 'to-team'), __('team', 'to-team'), __('subtitle', 'to-team')],
-            isActive: (blockAttributes, variationAttributes) => {
-                return blockAttributes.className === variationAttributes.className;
-            },
-            attributes: {
-                metadata: { name: __('Team Tagline', 'to-team') },
-                className: 'lsx-tagline-wrapper',
-                layout: { type: 'flex', flexWrap: 'nowrap', verticalAlignment: 'top' },
-            },
-            innerBlocks: [
-                [
-                    'core/group',
-                    { layout: { type: 'flex', flexWrap: 'nowrap', verticalAlignment: 'middle' } },
-                    [['lsx-tour-operator/icons', { iconType: 'solid', iconName: 'taglineIcon' }]],
-                ],
-                [
-                    'core/group',
-                    { layout: { type: 'flex', flexWrap: 'nowrap' } },
-                    [
-                        [
-                            'core/paragraph',
-                            {
-                                metadata: {
-                                    bindings: {
-                                        content: { source: 'lsx/post-meta', args: { key: 'tagline' } },
-                                    },
-                                },
-                                prefix: __('Tagline:', 'to-team'),
-                                prefixBold: true,
-                            },
-                        ],
-                    ],
-                ],
-            ],
-            example: {
-                innerBlocks: [
-                    {
-                        name: 'core/group',
-                        attributes: { layout: { type: 'flex', flexWrap: 'nowrap', verticalAlignment: 'middle' } },
-                        innerBlocks: [
-                            { name: 'lsx-tour-operator/icons', attributes: { iconType: 'solid', iconName: 'taglineIcon' } },
-                            { name: 'core/paragraph', attributes: { content: '<strong>' + __('Tagline: ', 'to-team') + '</strong>' + __('Expert safari guide with 20 years experience', 'to-team') } },
-                        ],
-                    },
-                ],
-            },
-        });
-    };
-
-    const conditionalRegister = registerForPostTypesAndTemplates(['team'], ['team'], registerTeamTaglineVariation);
-    conditionalRegister();
-});
